refactor: adjust animal type service into its own package
