+++ src/main/java/com/jpcchaves/adotar/application/service/animaltype/AnimalTypeServiceImpl.java
@@ -1,12 +1,14 @@
-package com.jpcchaves.adotar.application.service.impl.v1;
+package com.jpcchaves.adotar.application.service.animaltype;
 
 import com.jpcchaves.adotar.application.dto.pet.AnimalTypeMinDto;
-import com.jpcchaves.adotar.application.service.usecases.AnimalTypeService;
+import com.jpcchaves.adotar.application.service.animaltype.contracts.AnimalTypeService;
 import com.jpcchaves.adotar.application.utils.mapper.MapperUtils;
 import com.jpcchaves.adotar.domain.exception.ResourceNotFoundException;
 import com.jpcchaves.adotar.domain.model.AnimalType;
 import com.jpcchaves.adotar.infra.repository.AnimalTypeRepository;
+
 import java.util.List;
+
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,7 +17,8 @@ public class AnimalTypeServiceImpl implements AnimalTypeService {
   private final MapperUtils mapperUtils;
 
   public AnimalTypeServiceImpl(
-      AnimalTypeRepository animalTypeRepository, MapperUtils mapperUtils) {
+      AnimalTypeRepository animalTypeRepository,
+      MapperUtils mapperUtils) {
     this.animalTypeRepository = animalTypeRepository;
     this.mapperUtils = mapperUtils;
   }
