+++ src/main/java/com/jpcchaves/adotar/application/service/animaltype/contracts/AnimalTypeService.java
@@ -1,6 +1,7 @@
-package com.jpcchaves.adotar.application.service.usecases;
+package com.jpcchaves.adotar.application.service.animaltype.contracts;
 
 import com.jpcchaves.adotar.application.dto.pet.AnimalTypeMinDto;
+
 import java.util.List;
 
 public interface AnimalTypeService {
